challenge/requirements/base.txt
# Requeriments base.

django==3.1.3
djangorestframework==3.12.1
django-environ==0.4.5
django-filter==2.4.0
